+++ arkham/src/main/java/co/com/cesde/arkham/domain/dto/appointment/AppointmentListRecord.java
@@ -8,7 +8,13 @@ import co.com.cesde.arkham.domain.dto.user.UserListRecord;
 import java.time.LocalDate;
 import java.time.LocalTime;
 
-public record AppointmentListRecord(Integer appointmentId, LocalTime startTime, LocalTime endTime, LocalDate appointmentDate, PropertyListRecord propertyListRecord, UserListRecord userListRecord, ClientListRecord clientListRecord) {
+public record AppointmentListRecord(Long appointmentId,
+                                    LocalTime startTime,
+                                    LocalTime endTime,
+                                    LocalDate appointmentDate,
+                                    PropertyListRecord propertyListRecord,
+                                    UserListRecord userListRecord,
+                                    ClientListRecord clientListRecord) {
 
     public AppointmentListRecord(Appointment appointment){
         this(appointment.getAppointmentId(),
