+++ arkham/src/main/java/co/com/cesde/arkham/domain/dto/property/PropertyRegisterRecord.java
@@ -9,7 +9,7 @@ public record PropertyRegisterRecord(
         @NotNull
         Boolean free,
         @NotNull
-        Integer ownerId,
+        Long ownerId,
         @NotBlank
         String offer,
         @NotBlank
